code cleanup

+++ src/main/kotlin/com/github/su23/consul/app/leader/elector/config/ClusterConfiguration.kt
@@ -9,10 +9,10 @@ data class ClusterConfiguration(
 ) {
     companion object {
         data class Consul(val host: String = "localhost", val port: Int = 8500)
-        data class Session(val ttl: Int = 15, val refresh: Int = 7)
+        data class Session(val ttl: Long = 15, val refresh: Long = 7)
         data class Election(
-            val frequency: Int = 10,
-            val decay: Int = 5,
+            val frequency: Long = 10,
+            val decay: Long = 5,
             val envelopeTemplate: String = "services/%s/leader"
         )
     }
